+++ src/Candoumbe.Pipelines/Components/NuGet/IPushNugetPackages.cs
@@ -4,6 +4,7 @@ using Nuke.Common.IO;
 using Nuke.Common.Tooling;
 using Nuke.Common.Tools.DotNet;
 
+using System;
 using System.Collections.Generic;
 using System.Linq;
 
@@ -17,6 +18,10 @@ namespace Candoumbe.Pipelines.Components.NuGet;
 /// <summary>
 /// Marks a pipeline that can push NuGet packages to various repositories
 /// </summary>
+/// <remarks>
+/// By default, this component publishes nuget packages specified by <see cref="PublishPackageFiles"/> to repositories defined by <see cref="PublishConfigurations"/>.
+/// This behavior can be overriden by explicitely specifying the <see cref="ConfigName"/>.
+/// </remarks>
 public interface IPushNugetPackages : IPack
 {
     /// <summary>
@@ -25,16 +30,24 @@ public interface IPushNugetPackages : IPack
     IEnumerable<AbsolutePath> PublishPackageFiles => PackagesDirectory.GlobFiles("*.nupkg", "*.snupkg");
 
     /// <summary>
+    /// Explicitely
+    /// </summary>
+    [Parameter("Defines the name of the configuration to use to publish packages.")]
+    public string ConfigName => TryGetValue(() => ConfigName)?.Trim();
+
+    /// <summary>
     /// Publish all <see cref="PublishPackageFiles"/> to <see cref="PublishConfigurations"/>.
     /// </summary>
     public Target Publish => _ => _
-        .Description($"Published packages (*.nupkg and *.snupkg) to the destination server set using {nameof(PublishConfigurations)} settings ")
+        .Description($"Published packages (*.nupkg and *.snupkg) to the destination server set using either {nameof(PublishConfigurations)} settings or the configuration {nameof(ConfigName)} configuration ")
         .Consumes(Pack, ArtifactsDirectory / "*.nupkg", ArtifactsDirectory / "*.snupkg")
         .DependsOn(Pack)
-        .OnlyWhenDynamic(() => PublishConfigurations.AtLeastOnce(config => config.CanBeUsed()))
-        .WhenNotNull(this as IHaveGitRepository,
+        .OnlyWhenDynamic(() => (!string.IsNullOrWhiteSpace(ConfigName) && PublishConfigurations.Once(config => config.Name == ConfigName))
+                                || PublishConfigurations.AtLeastOnce(config => config.CanBeUsed()))
+        .WhenNotNull(this.As<IHaveGitRepository>(),
                      (_, repository) => _.Requires(() => GitHasCleanWorkingCopy())
-                                         .OnlyWhenDynamic(() => repository.GitRepository.IsOnMainOrMasterBranch()
+                                         .OnlyWhenDynamic(() => IsLocalBuild
+                                                                || repository.GitRepository.IsOnMainOrMasterBranch()
                                                                 || repository.GitRepository.IsOnReleaseBranch()
                                                                 || repository.GitRepository.IsOnHotfixBranch()))
         .Requires(() => Configuration.Equals(Configuration.Release))
@@ -43,30 +56,50 @@ public interface IPushNugetPackages : IPack
             int numberOfPackages = PublishPackageFiles.TryGetNonEnumeratedCount(out numberOfPackages)
                 ? numberOfPackages
                 : PublishPackageFiles.Count();
+            DotNetNuGetPushSettings pushSettings = new();
+            pushSettings.Apply(PublishSettingsBase)
+                        .Apply(PublishSettings);
 
-            PublishConfigurations.ForEach(config =>
+            if (string.IsNullOrWhiteSpace(ConfigName))
             {
-                if (config.CanBeUsed())
+                PublishConfigurations.ForEach(config =>
                 {
-                    Information("{PackageCount} package(s) will be published to {SourceName}({SourceUrl})", numberOfPackages, config.Name, config.Source);
-                }
-                else
-                {
-                    Warning("{PackageCount} package(s) will not be published to {SourceName}({SourceUrl})", numberOfPackages, config.Name, config.Source);
-                }
-            });
+                    if (config.CanBeUsed())
+                    {
+                        Information("{PackageCount} package(s) will be published to {SourceName}({SourceUrl})", numberOfPackages, config.Name, config.Source);
+                    }
+                    else
+                    {
+                        Warning("{PackageCount} package(s) will not be published to {SourceName}({SourceUrl})", numberOfPackages, config.Name, config.Source);
+                    }
+                });
 
-            DotNetNuGetPush(s => s.Apply(PublishSettingsBase)
-                                  .Apply(PublishSettings)
-                                  .CombineWith(PublishPackageFiles,
-                                               (_, file) => _.SetTargetPath(file))
-                                                             .Apply(PackagePublishSettings)
-                                                             .CombineWith(PublishConfigurations,
-                                                                          (setting, config) => setting.When(config.CanBeUsed(),
-                                                                                                            _ => _.SetApiKey(config.Key)
-                                                                                                                  .SetSource(config.Source.ToString()))),
-                                                  degreeOfParallelism: PushDegreeOfParallelism,
-                                                  completeOnFailure: PushCompleteOnFailure);
+                DotNetNuGetPush(s => s.Apply(PublishSettingsBase)
+                                      .Apply(PublishSettings)
+                                      .CombineWith(PublishPackageFiles,
+                                                   (_, file) => _.SetTargetPath(file))
+                                                                 .Apply(PackagePublishSettings)
+                                                                 .CombineWith(PublishConfigurations,
+                                                                              (setting, config) => setting.When(config.CanBeUsed(),
+                                                                                                                _ => _.SetApiKey(config.Key)
+                                                                                                                      .SetSource(config.Source))),
+                                                      degreeOfParallelism: PushDegreeOfParallelism,
+                                                      completeOnFailure: PushCompleteOnFailure);
+            }
+            else
+            {
+                PushNugetPackageConfiguration publishConfiguration = PublishConfigurations.Single(config => string.Equals(config.Name, ConfigName, System.StringComparison.OrdinalIgnoreCase));
+                DotNetNuGetPush(s => s.Apply(PublishSettingsBase)
+                                      .Apply(PublishSettings)
+                                      .When(publishConfiguration.CanBeUsed(),
+                                            _ => _.SetApiKey(publishConfiguration.Key)
+                                                  .SetSource(publishConfiguration.Source))
+                                      .CombineWith(PublishPackageFiles,
+                                                   (_, file) => _.SetTargetPath(file))
+                                                                 .Apply(PackagePublishSettings),
+                                degreeOfParallelism: PushDegreeOfParallelism,
+                                completeOnFailure: PushCompleteOnFailure);
+            }
 
             ReportSummary(summary =>
             {